+++ source/include/acstruct.h
@@ -1,7 +1,7 @@
 /******************************************************************************
  *
  * Name: acstruct.h - Internal structs
- *       $Revision: 1.41 $
+ *       $Revision: 1.42 $
  *
  *****************************************************************************/
 
@@ -161,7 +161,7 @@ typedef struct acpi_walk_state
 
     ACPI_PARSE_STATE                ParserState;        /* Current state of parser */
     UINT32                          PrevArgTypes;
-    UINT8                           ArgCount;           /* push for fixed or var args */
+    UINT32                          ArgCount;           /* push for fixed or var args */
 
     struct acpi_namespace_node      Arguments[ACPI_METHOD_NUM_ARGS];        /* Control method arguments */
     struct acpi_namespace_node      LocalVariables[ACPI_METHOD_NUM_LOCALS]; /* Control method locals */
